+++ .vscode/launch.json
@@ -5,7 +5,7 @@
       "name": "Debug test",
       "type": "node",
       "request": "launch",
-      "program": "${workspaceRoot}/scripts/debug-test.js",
+      "program": "${workspaceRoot}/scripts/executors/debug-test.js",
       "cwd": "${fileDirname}",
       "stopOnEntry": false,
       "args": ["-i", "--watch"],
@@ -22,7 +22,7 @@
       "name": "Debug current open test",
       "type": "node",
       "request": "launch",
-      "program": "${workspaceRoot}/scripts/debug-test.js",
+      "program": "${workspaceRoot}/scripts/executors/debug-test.js",
       "cwd": "${fileDirname}",
       "stopOnEntry": false,
       "args": ["-i", "--testPathPattern=\\b${fileBasenameNoExtension}", "--watch"],
@@ -39,7 +39,7 @@
       "name": "Debug current open test (v0)",
       "type": "node",
       "request": "launch",
-      "program": "${workspaceRoot}/scripts/debug-test.js",
+      "program": "${workspaceRoot}/scripts/executors/debug-test.js",
       "cwd": "${fileDirname}",
       "stopOnEntry": false,
       "args": ["-i", "--testPathPattern=\\b${fileBasenameNoExtension}", "--watch"],
