+++ package.json
@@ -157,6 +157,7 @@
     "@types/node-fetch": "2.5.7",
     "@types/prettier": "2.2.3",
     "@types/progress": "2.0.5",
+    "@types/prompts": "2.4.1",
     "@types/react": "17.0.44",
     "@types/react-dom": "17.0.15",
     "@types/react-is": "17.0.3",
@@ -219,8 +220,8 @@
     "express": "4.17.1",
     "extract-comments": "1.1.0",
     "file-loader": "6.2.0",
-    "fork-ts-checker-webpack-plugin": "6.1.0",
     "find-free-port": "2.0.0",
+    "fork-ts-checker-webpack-plugin": "6.1.0",
     "fs-extra": "8.1.0",
     "geckodriver": "3.0.2",
     "glob": "7.2.0",
@@ -262,6 +263,7 @@
     "prettier": "2.2.1",
     "pretty-bytes": "5.6.0",
     "progress": "2.0.3",
+    "prompts": "2.4.2",
     "puppeteer": "14.4.0",
     "raw-loader": "4.0.2",
     "react": "17.0.2",
