+++ packages/react/package.json
@@ -20,7 +20,7 @@
     "build-storybook": "cross-env NODE_OPTIONS=--max-old-space-size=3072 just-scripts storybook:build",
     "clean": "just-scripts clean",
     "code-style": "just-scripts code-style",
-    "codepen": "node ../../scripts/local-codepen.js",
+    "codepen": "node ../../scripts/executors/local-codepen.js",
     "e2e": "yarn workspace @fluentui/react-examples cypress run --component",
     "e2e:local": "yarn workspace @fluentui/react-examples cypress open --component",
     "just": "just-scripts",
