+++ scripts/executors/format.js
@@ -1,11 +1,9 @@
-// @ts-check
-
 // This script is not meant as a general-purpose replacement for prettier.
 // It only handles certain cases that prettier doesn't support as well/at all natively.
 
 const { execSync } = require('child_process');
-const { runPrettier, runPrettierForFolder } = require('./prettier');
-const { findGitRoot } = require('./monorepo');
+const { runPrettier, runPrettierForFolder } = require('../prettier');
+const { findGitRoot } = require('../monorepo');
 
 const parsedArgs = parseArgs();
 const root = findGitRoot();
