+++ scripts/package.json
@@ -36,7 +36,6 @@
     "p-queue": "^6.0.2",
     "path-browserify": "^1.0.1",
     "plop": "^2.6.0",
-    "prompts": "^2.3.0",
     "react-hot-loader": "^4.13.0",
     "react-vis": "^1.11.6",
     "read-pkg-up": "^7.0.1",
