+++ tools/generators/migrate-v8-pkg/index.spec.ts
@@ -124,7 +124,7 @@ function setupDummyPackage(
         'build-storybook': 'cross-env NODE_OPTIONS=--max-old-space-size=3072 just-scripts storybook:build',
         clean: 'just-scripts clean',
         'code-style': 'just-scripts code-style',
-        codepen: 'node ../../scripts/local-codepen.js',
+        codepen: 'node ../../scripts/executors/local-codepen.js',
         e2e: 'yarn workspace @fluentui/react-examples e2e --package react',
         just: 'just-scripts',
         lint: 'just-scripts lint',
